set rpcs from registry + refactor
